Moved a couple of helper methods into a shared helper.
