+++ assets/components/contributionSelection/helpers.js
@@ -2,7 +2,11 @@
 
 // ----- Imports ----- //
 
-import { config as contributionConfig } from 'helpers/contributions';
+import {
+  config as contributionConfig,
+  getOneOffSpokenName,
+  getSpokenType,
+} from 'helpers/contributions';
 import { spokenCurrencies } from 'helpers/internationalisation/currency';
 
 import type { Contrib as ContributionType } from 'helpers/contributions';
@@ -95,25 +99,6 @@ const amounts = {
 
 function getOneOffName(country: IsoCountry) {
   return country === 'US' ? 'One-time' : 'One-off';
-}
-
-function getOneOffSpokenName(country: IsoCountry) {
-  return country === 'US' ? 'one time' : 'one off';
-}
-
-function getSpokenType(
-  contributionType: ContributionType,
-  country: IsoCountry,
-) {
-
-  if (contributionType === 'ONE_OFF') {
-    return getOneOffSpokenName(country);
-  } else if (contributionType === 'MONTHLY') {
-    return 'monthly';
-  }
-
-  return 'annual';
-
 }
 
 function getContributionTypeClassName(contributionType: ContributionType) {
